[plugin] bookshelf: add finances micromodule
